zaf: only iterate over instances if class has get_config
The way I understand the old code, it would check if the extension class had
the get_config attribute, but it would do this for every instance of the class.
I figured there are two ways this code was meant to work:

* for every instance, check if the *instance* has 'get_config', or
* check if the class has 'get_config' before iterating the instances.

I chose to go with the latter implementation.